homework5/requirements.txt
docopt >= 0.6.2, < 0.7
envoy >= 0.0.3
github3.py >= 0.9.5
importlib >= 1.0.4
numpy >= 1.13.3, < 1.14
ordereddict >= 1.1
python-dateutil >= 2.2, < 2.3